Changes, Roadblocks, etc.
need to figure out how to check if fishy is eating poison or food.
Thought I should use an interface cause of common isPoison() method,
but in finding the distance, it was hard to check the distance to a
nourishment type. so i had to make two separate distance methods which
i find inefficient.

started to write in which testers I want to have.