what even is this error?
just made one class for the Nourishments. On a coin toss determines
whether food drawn should be red or yellow (poison, food). If coin toss
= true, food drawn. if coin toss= false, poison drawn. Coin toss method
helps me also tell whether or not nourishment is food/poison! This
makes a lot more sense now.

Now i just have to figure out the error that says, “duplicate class:
game1.Nourishment.food”. I’m very confused as to why that is a mistake.
It should not be anymore.